+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sveltejs/kit
 
+## 1.0.0-next.78
+
+### Patch Changes
+
+- 6e27880: Move server-side fetch to adapters instead of build step
+- 61d7fa0: Better error logging
+- 041b706: Implement layout resets
+- 148819a: Use latest vite-plugin-svelte
+- 9d54eed: Make sveltekit:prefetch a noop if <a> has no href
+
 ## 1.0.0-next.77
 
 ### Patch Changes